Add some rudimentary support for unauthenticated resources
The plugin currently assumes that a secret and access keyfile WILL be provided in the xrootd configuration.
However, some origins may be exporting public buckets that don't require any authentication. In these cases,
it appears that providing an `Authorization` header causes problems, so the plugin should detect and handle
the case when no access/secret keys are provided.