Add rudimentary plugin support for regular HTTP server
The S3 plugin already had a lot of the machinery necessary to interface with regular
HTTP servers, so I duplicated all the S3 Files and made the changes needed to
interface with regular HTTP. I started going back to deduplicate the code between the
two plugins, but a lot of that work remains.

Also different in this commit is the CMakeLists.txt now builds two objects -- one for
the S3 plugin (as before) and one for the HTTP server plugin.