Fix issue with host header and canonicalURI
There were actually two issues that broke authenticated buckets. The first
was that we weren't properly setting the host header. The second was that
we weren't generating a valid canonical URI that gets signed and sent as
a part of the Authorization header. The solution to both of these was to
parse the hostUrl more carefully, and in only one spot.