Add test and gtest submodule to run unit tests
This testing framework is largely cloned from the [scitokens-cpp](https://github.com/scitokens/scitokens-cpp)
library. We may need to modify a few things to best suit our needs, but at this
point it should be running the minimal unit tests we have and proving that
the source code builds.

Right now this only builds on ubuntu. We should update this to build/test on
MacOS soon.