Improve handling of errors for fstat failures
The HTTP server will do a HttpFileSystem::Stat (ignoring the result) followed
by a HttpFile::Open.  Since the HttpFile::Open was a no-op, the fact a
file didn't exist (or was otherwise broken) wasn't recorded, a HTTP 200 was
always returned (and _then_ there was a subsequent failure when transferring).

This change causes a stat to always be done on file-open, causing the correct
HTTP status code to be sent.

Additionally, it improves the error code handling for both HTTP and S3 stat's.