Fix some cmake glitches
Do not overwrite the original values of CMAKE_SHARED_LINKER_FLAGS and
CMAKE_MODULE_LINKER_FLAGS. The original values contain the content of
the LDFLAGS environment variable. If these variables are overwritten
instead of extended, the content of LDFLAGS will be ignored.

Use the module name when calling find_package, not the target name
defined in the module.