Avoid unpausing from a separate thread
We want to always unpause a given operation from the same thread that
is handling it.  If a separate thread can pick up the operation, there
is a race condition where both the original one and new one are operating
on the same `CURL *` handle at the same time, resulting in an observed
segfault.

This commit introduces a separate "unpause" queue for each curl worker;
this queue is notified by the parent when there is additional data
available.