Time out in-progress transfers
If an entire part is a single libcurl operation, a client that starts
writing - but then leaves for an extended period of time - will leave
dangling references inside libcurl (eventually exhausting the number
of allowable transfers).

This commit adds a background thread for S3File that will go through
all pending uploads and check to see if they are still live; if not,
then it'll timeout the operation.