Add buffering + prefetch scheme to S3 reads
This keeps a 2-entry cache inside the S3File object.  These cache entries
round up small reads and allow prefetching in the case of linear reads.

The goal is to establish pipelining when doing linear reads, allowing
data to be read to the S3File cache entry before it is requested.