Improve handling of empty objects
The prior fix for handling of empty object names (such as when you
are listing the root of the bucket) depended on counting `/` characters
(which may be inconsistent if the paths aren't normalized) and didn't
cover the `S3FileSystem::Create` and `S3File::Open` methods.

This pushes the fix into the `parsePath` method and adds unit test
coverage to watch for regressions.