Fix potential deadlock
The mutex protecting the state variables of the cache was held when
the read request was submitted to the queue.  This worked well when
the queue was not at capacity -- but failed when the submission to
the queue blocked.  This is because the mutex is needed by the
callback function to finish operations.

The fix was to drop the mutex while we're entering the queue and to
acquire it again afterward.