Turn HTTP 503 into an error
Previously, it looked like the HTTP 503 handling could result in a
retry.  This was a lie - it was code remnants from before there was
the libcurl multihandle work.

Instead, the "retry" handling caused the response to be marked as
successful with the error message in the body used as the successful
response.  This resulted in corrupt garbage being returned to the
client.

This removes the `CurlResult` enum (which was actually unused) to
prevent future confusion.