PATH_MAX is not portable
Not all systems define PATH_MAX and/or MAXPATHLEN. The proper fix is
to use dynamic allocation and allocate the number of bytes needed.
This is a bit of work, and it has not been done for the main xrootd
code. The xrootd code has a hack and defines MAXPATHLEN in
<XrdSys/XrdSysPlatform.hh> on system that don't have it in the system
headers. Use this hack in the POSC plugin too.