Fix broken file access mode detection
The access mode values for files are not the same for all platforms.
The detection of a read only request here assumes that O_RDONLY is 0,
which is not mandated by the POSIX standaed.

Linux:
  #define O_ACCMODE       00000003
  #define O_RDONLY        00000000
  #define O_WRONLY        00000001
  #define O_RDWR          00000002

But this is also allowed:
  #define O_RDONLY      0x0001  /* Open read-only.  */
  #define O_WRONLY      0x0002  /* Open write-only.  */
  #define O_RDWR        (O_RDONLY|O_WRONLY) /* Open for reading and writing. */
  #define O_ACCMODE     O_RDWR  /* Mask for file access modes.  */

This commit fixes the detection so that it is done in the intended way
that works everywhere.