Implement atomic writes
There are two parts to this commit, the first being that we are spoofing
creates because we are handling it in the call to Open. The second part
is using the `oss.asize` to give a hint on how big the upload is so that
we can ensure that the file is only commited when all expected bytes are
on disk.